Basic stubs and initial pseudo-implementation
The basic muLJF system with the extended encoding is complete insofar
as dependencies and fundamental syntax goes. Some important questions
remain to declare a first version of the kernel.

Naming conventions follow those used by the muFa checker, whether
tweaks are desirable remains unaddressed for now. A trivial and likely
wrong in many places FPC is stubbed to base some tests on it.