Integrated generic fixed points and tests
A basic implementation of fixed points with arbitrary argument lists
reuses the existing one-place implementation. The single type i is
augmented with constructors for special lists of arguments as a quick
implementation that does NOT prevent misuse due to type sharing.
Presently, this is the fastest and cleanest way to experiment.

In addition, several tests are added, including tentative specification
of fixed point predicates and pointers to potential test batteries.
The fundamental test battery for the individual rules of inference in
the kernel does not depend on nonlogical constants and is ordered in
such a way that the tests of a rule, represented by its unique clerk/
expert, may only use verified rules to conclude its derivation.X