FPC for the totality of list concatenation
Inductive proofs are generated for all standing examples on the append
predicate. The main novelties are (trivial) missing store and release
clerks and experts, and an extension of the indexing mechanism for the
avoidance of looping derivations. A single "bucket" of formulas risks
deadlock in at least certain derivations that do not discriminate
which stored formulas have been chosen and which have not, also when
competing with stored formulas on the right.

The issue is perhaps a bit more subtle and may reoccur if handled
without care. The tweak adds counters to all affected certificates to
keep track of the current state of affairs, which is neither too
modular nor, possibly, bulletproof. But since we are in "move fast and
break things" mode, this is tolerable for the time being.