New (failing) examples on append
The determinacy of append evidences in particular the necessity to
extend the kernel. At a given point in the derivation, a sequent of
the following shape is encountered (in muLJF notation):
   (append nil p l')* |- p = l'

The only alternative for the fixed point was freezing, since no
further induction was planned. At this point we are confronted by a
case of synchronous eigenvariable unification, which cannot be
resolved.

Moreover, the synchronous phase that follows can only be followed by
an infinite loop (broken backtracking) which selects the equality
ad aeternum, even though it could be inferred that there is nothing
to be done. This is a weakness that may need addressing in future.

In order to deal with this situation, unfolding will be required.