Implementation of unfolding rules
The kernel is extended with rules for performing unfolding on both
least and greatest fixed points. It may be interesting to consider
the order of application of all three asynchronous operations on
fixed points (unfolding, (co)induction, freezing).

Additionally, an error in the encoding of the freezing rule on the
right has been corrected.