Restricted unfolding in append FPC
Append certificates feature new placeholders for counters of allowed
unfolding operations. It is not entirely clean, but it should suffice
at this early stage.

Based on this, existing tests are preserved and determinacy of append
shows progress. The following runtime error on the original induction
clause BSx is encountered.

This formula embeds S with its explicit signature of universal
quantifiers as the new recursive predicate. The goal of the formula is
of the form k = l, for k, l part of the universally quantified local
signature. Eventually, the focus sets on this Sx and logical variables
replace the bound variables, therefore K = L, indeed the end goal.
Treatment of the nested implications must result in a sequent:
   ..., [K = L] |- ...

The following runtime error results:
   Logic variable encountered on the left: H85.

If this situation is indeed correct, and given that by design Bedwyr
will not reschedule lines of inference, order of subgoals may be
split and guided by specialized clerks and experts to account for
this unfortunate limitation, in hopes that other lines of inquiry
will instantiate the logic variables sufficiently to continue.