Specular left implication with append tests
A specular rule for left implication which inverts the order in which
subgoals are handled has been added to the kernel.

For reasons pertaining to the nature of the goals being proved (as
expounded in my previous commit), this inversion becomes the favored
option and replaces the standard left implication rule for all intents
and purposes.

All examples up to determinacy of append continue to pass, also under
full induction. Said determinacy property is now provable with one
unfolding step.