Test cases for addition predicate
The addition predicate, plus, and its associated properties are given
and verified. Taking append as a template and reusing its FPC, those
properties given in the Abella tutorial and the typical ones that were
tested for append are defined analogously. Everything works as expected.

Ample opportunities for refactoring and sharing appear. Target
properties (and their invariants and certificates) appear from our
as of yet limited evidence to be templates that can be reused with
significant flexibility.