Stubs for automatic induction
The approach being followed is a straightforward simplification of
standard induction, where information provided explicitly will be
maintained by the kernel.

I do not want to involve certificates in the partial or total guidance
of this process, since an incorrect specification would risk the
soundness of the induction. For this reason, the language of sequences
will have to be enriched, at the very least to be able to hold the
current eigenvariable signature (details are left implicit in the
papers, so proceed carefully to ensure its correct maintenance).