Sequent signatures for derivation of invariants
Immediate invariants for induction are derived automatically. These
depend on the contents of the sequent to which the induction rule is
applied. In order to build the signature appropriately, a new area
Sigma is added and maintained for all derivation rules.

Integrating this functionality with the helper predicates for
invariant derivation (still in a less-than-perfect embryonary state)
causes unification errors due to the logic variable-like nature of the
contents of Sigma. Once corrected, the new functions should be able
to integrate seamlessly with the rest of the kernel.

Note that only the examples for the plus predicate have been extended
to comply with the new signature; all other batteries are pending
modification (upon completion of the core functionality).