Improved invariant derivation
Immediate invariants are correctly, though not optimally, constructed.
There remains the issue of sequencialization of code, which may be
addressed by signature files for the kernel and each battery of tests.

The problem is as follows: the new implementation depends on an
exhaustive definition of the copy_i predicate, which depends on the
structure of type i, itself (partially) defined by the user. This
information must be declared before the kernel is included, and the
kernel itself is required before the sequents to be proved are
defined.