Partial reimplementation of immediate invariants
Following Dale's suggestion, a new method for computing the invariants
substitutes the previous, imperfect iteration of the procedure. This
seems to work nicely in most cases, though a need to supply an
exhaustive predicate copy_i forces us to split each test battery in a
signature that can furnish this information to the kernel, and the
tests proper.

For most examples this enables immediate induction to proceed, albeit
seemingly with potential backtracking. The associativity property proves
its weaknesses and fails to complete as expected, appearing to loop
instead. I believe the problems stem from a naive approach to
representing signatures, a mildly thorny issue that I hope to banish
in the interest of simplicity.