Reformulation of associativity property
All examples on plus are working (even though Sigma is still a mess,
but since we only use it once at the very beginning, it isn't giving
trouble... but it is hacky and must be remedied).

The culprit here was the formulation of the associativity theorem and
in particular the canonical order of its two implicational hypotheses.
An induction on the first formula, say (plus A B AB) succeeds, whereas
one on (plus AB C ABC) does not, or not under these constraints.

Because we always process the head of the list of formulas and the last
formula to be induced there was precisely the second fixed point,
induction was proceeding on the wrong clause (the certificate is, at
least for now, too blind to discriminate what clause it has to aim for,
but this greedy quality may be one of its strong points).

Permuting both implicational hypotheses has no proof-theoretical effect
and yields the sequent that we expected to be able to proof.

Good debugging and symbolic manipulation tools would have made this so
much more straightforward.