Restructuring of administrative certificate
Clerks and experts are organized in akin groups.

Some helper predicates are declared and used to refactor those clerks
and experts that manipulate indexes, and do so for each separate type
of certificate.

On a related issue, said certificates are completed with missing
(repeated) specifications for immediate induction (autoinduce), which
should have had no effect until now. Nonetheless, an evident risk of
not being sufficiently DRY.

A check is added to prevent some race conditions when deciding on the
left.

To avoid exotic backtracking issues, all predicates have been defined,
with those not meant to be used instructed to always fail.

Tentatively, a restriction on freezing on the left has been added to
operate exclusively on start certificates.

Minor comments and points of interest added in square brackets.

This is a rather messy commit that should be split in several thematic
ones. Such is life.