Scaffolded example of use of lemmas
A modified initR rule tries to apply the totality of addition as a lemma
to prove everything else in the universe (and succeeds for a modified
statement of the totality of multiplication which includes an explicit
repetition of its hypothesis in order to make it available as a frozen
atom).

The general approach is sane and being so lazy about application seems a
good thing, even if there are questions concerning its technical
implementation in the general case. As Dale favors decideL, we will try
to transplant this idea and pray we don't run into the looping behaviors
that still haunt the kernel.