Debugging module for kernel
A custom debugging module is added for the kernel and related
submodules. The idea is to provide type-specific pretty printers and
spy predicates integrating these printers for those kernel predicates
deserving special attention.

This is an initial implementation of some of the basic functions. It
(ab)uses the explicit notion of a signature of fixed points for which
names are known, and used to produce readable versions on screen.
However, this signature is given explicitly, which would force to
modify all involved predicates in a more than suboptimal way, just in
case we want to trace them occasionally.

To improve on this, we will define a "debugging interface" in the form
of a predicate indexing these fixed points, and the implementation of
this predicate given by interested users will be called to access the
index while the kernel remains unchanged.