Lemmas in kernel (part 1 of ?)
A new "global" decide rule on the left hand side is used to introduce
lemmas from a new sequent zone Lambda containing named theorem-like
formulas, with an immediate left decide on the new introduction.
Compare with the traditional "local" decide rule.

Some limited indexing is added to try to avoid read duplicates, but at
this point I am unsure whether this is necessary or how it will end up
being used.

The ordering of decide rules, which is tweaked in this commit, remains
questionable.

The totality of the multiplication operation has been adapted to play
by the new rules, but is not functional yet for several reasons that
will be addressed shortly.