Totality of multiplication
The modifications to the kernel sufficed to retain all its desirable
behaviors while affording enough flexibility to make effective use of
lemmas for the reference example. What was needed was an adequate
strenghtening of the totality template, in particular the introduction
of a composite goal with a typing judgment after the main conclusion
to be derived.

It turns out merging the newest features in master was not strictly
necessary, even though it does not hurt to have them.