Limit number of async-sync phases
Kernel extended with a limited search bipole depth (using natural
indexes to do this in a quick and dirty way). This is a necessary but
not sufficient condition to avoid nonterminating search. It should
handle the problem of choice well enough, but is not strong enough to
deal with unbounded recursion on free variables during the synchronous
phase.

All examples work except those on the times operation, which present
the aforementioned difficulties and call for further limits on
the synchronous phase, and specific to it.