Limit number of synchronous unfoldings
Bound added to (globally) limit the number of (currently right)
unfoldings in a proof, with default values made to match all passing
testsets.

Times tests that should fail do not, even though most sources of
infinite recursion have been quelled. There remains the same problem
that was found with Delta: a lemma is applicable an arbitrary number
of times, and the Bedwyr system does not recognize when it has
exhausted its options.

Pending points:

 - Most importantly, address issue of repetitive choice of lemmas.

 - Limit synchronous unfoldings on the right.

 - Consider asynchronous unfoldings (not immediately relevant for us).

 - Reset unfolding count for each synchronous phase.