Bounded unfoldings split and ported to FPC
Unfolding bounts have been split in a global limit for asynchronous
unfoldings (on the left) and a per-phase limit for synchronous
unfoldings (on the right). Furthermore, the kernel has been rolled
back to its "boundless" form and its bounding facilities ported to
the administrative FPC.

All examples have been modified under the new discipline. Ways to
manipulate indexes of interest in a compact way (and to encode them
in one such readable fashion) would be desirable.