Search depth moved from kernel into FPC
Following our initiative to extirpate excess logical tissue out of the
kernel, we roll back to HEAD~4 and implement depth search bounds as a
nonessential component in the administrative FPC.

All examples have been readjusted. Failure to specify a sufficient
depth may result in looping behavior instead of the desired
termination of proof search. This fallback case is common and can be
observed e.g. in append_total. Decide rules, in particular those on
the right, are possibly the main offenders, since we do not keep track
of what we have done before (there is only one formula, after all).

Note that it may be best to specify depth at the top level alone and
propagate it whenever a new certificate is extracted or created in the
FPC. This is, in fact, probably the intuition we have.