Deterministic asynchronous processing
A source of exponentially rich variations during backtracking is the
choice we enjoy during the asynchronous phase: given a list of formulas
on the left and a formula on the right, where do we start? Even if we
process the first formula of the left list, the issue remains.

The kernel has been modified to always process the left side first, and
the right side only when no formulas to be processed remain on the left
(i.e. we've "burned through" all of those). A similar strategy is used
in Zak and Dale's kernels.

The FPC still (re)defines the choice of whether to freeze or unfold a
formula on the left: previously, we would freeze when no unfolding
quota remained. But given our augmented control structure we may want to
freeze an atom also when we would be allowed to unfold it (for example,
because treatment differs across branches, keeping in mind that branching
retains the control structure). This can make unfolding inefficient if
bounds are not tight, and in fact dynamic bound growth better, but it
should not disturb correctness of the method, and it would allow
everything to fail rather than loop.

With this and minor adjustments in the order of template hypotheses (now
we don't need to worry about messing with this order to induce on the
right one: the new greedy processing rules make this matter trivial) we
preserve all working examples, except those involving lemmas, where
given the inner workings of the Bedwyr system a smarter way to cycle
through them will be needed.