Lemma indexing and simple implementation
The FPC has been extended with a counter for lemmas, and the definition
of lemma extended with an explicit index parameter, all this to avoid
looping reads and ensure the termination of any search.

There seems to be a legitimate point of choice between local and global
decideL[']Clerk (i.e., stored formulas on the left and lemmas).
Backtracking hinders the explicit ordering of stores before lemmas, and
the check has been (at least temporarily) dropped.

A minor kernel tweak reflects the explicit indexing.

All control structures in the examples have been adjusted and work.
Specifically, because of the order of operations, the lemma needs to be
applied twice in what I believe is a successive instance of an async
phase decision. A better alternative may be to reset lemma counters and
consider all lemmas for each decision in turn: smarter but slower,
because our system is not too clever.