Smarter lemma indexing
The administrative FPC is extended with a double lemma counter covering
the index of the last lemma to be considered from the list of lemmas,
assuming it exists, and incrementing index of the lemma to be selected
for the task. These two counters, LNext and LLast, should normally be
initialized with z and the index of the last declared lemma,
respectively. In many cases placeholders _ _ can be used.

Existing control structure functions are adjusted and especially those
in charge of LNext, which control the selection of a lemma to be applied
through decideLClerk' by a mixed recursive strategy:

 - The lemma corresponding to the current value of the LNext counter is
   selected and the global decide rule applied. This resets the counter
   to consider all declared lemmas at the next decision point.

 - The current lemma is ignored in favor of a selection on the remaining
   lemmas in the list, achieved by incrementing the counter and applying
   the selection procedure recursively.

All harnesses are fixed to work with the extended FPC. In particular,
totality of multiplication is proved with a single instance of the lemma
totality of addition, as was intended.