Commutativity of plus
The commutative property of addition can now be proved. To do this,
two previous results are needed as lemmas. The statement of the
property is extended, as is the certificate. Furthermore, the
property template now accepts an explicit list of lemmas.

To do: study the abstraction of the aforementioned list of lemmas,
and importantly reference complex lemmas in the list of previous
results to avoid their inlining (DRY up!).