A guided FPC extension for branching and lemmas
This simple extension to the administrative FPC allows the representation of
guidance trees and respects all existing examples that do not rely on
lemmas. For now branching instructions are limited to the asynchronous
disjunction, and do not rule out regular simple search. Each nested
certificate has access to its own constrained limits of computation.