FPC: freezeL works for all "search" certificates
It must be possible to freeze atoms at the very least whenever the
certificate allows a sort of unfocused search: unlike those certificates
involved in induction, which look for opportunities to apply their special
rule and do so with extreme prejudice (in principle, so as to rule out an
excessive number of choices, and barring more elegant/flexible treatment
involving positionals, delays, etc.).

Some minor formatting adjustments are also incorporated.