Branching certificates for plus and times
All current examples involving lemmas are given new branching certificates
that complete the work successfully.

Note that there is added efficiency in imposing sensible bounds for each
stage in the computation, in that these limits are not global and are, in
fact, much more local than hitherto, which reduces choice significantly.