Typing judgment for the augend
Working up to commutativity of multiplication. Here I experiment with
conventions to "publish" each significant formula in a dedicated fixed
point predicate so as to ease use of formulas in lemmas and other
expressions depending on previously defined formulas.

In addition, it is to be noted that a case analysis simplifies the
definition of nested certificates and their bounds, even if a huge,
blind certificate will always eventually get to the answer. There are
some gaps that prevent full systematization, especially regarding when
to unfold successor-like fixed points prior to freezing, for preparation
of future init rules on the right, and when computation on the right
is needed at the tips (of the kind performed by Abella's "search").

Nonetheless, it's coming together.