Bug: chained FPFs must bequest their global state
The state that transcends a sequence of certificates (in this case, at
the very least, the linear use of indexes of formulas stored in Gamma)
must be communicated to any succeeding certificate to preserve these
global restrictions.

Note that if certificate transitions may occur during the synchronous
phase, there are still fixes to be made. Whenever certificates are
nested and replaced, an adequate bequest must take be effected.

The exception here are currently the inductions, out of laziness and
given that the current model only considers their simple application
at the very beginning.