Symmetric associativity of addition via lemmas
A constructive method to prove the associativity of addition on the
right using previously proved properties. The certificate is given as
a linear sequence of lemma applications (with some potential
parallelism) which, despite signficant backtracking due to multiple
matches during choice of fixed points, ends clearly, predictably and
without problems.