Guided hypothetic reasoning (GHR), kernel changes
To support reasoning guided by hypotheses, e.g. informing the checker
which hypotheses to use to prove which lemma, some minimal changes to
the kernel seem in order.

The fundamental adaptation is indexing of frozen fixed points as the
essential (exclusive?) components of the list of hypotheses in lemmas.

The first step to do this is generalize the indexing scheme, so that
logic formulas may be given external naming rules; to this end, a new
type of indexes for formulas is introduced: boolidx. It must be noted
that, currently, there exists a mix in the kernel of several
incarnations of indexes, which must eventually be purged, as well as
the interfaces simplified.

With the addition of indexing to the frozen zone, all storage areas in
the sequents can be searched. In addition, the following changes are
made:

 - freezeLClerk gains an index output parameter to direct the key with
   which a fixed point will be stored in the frozen area.

 - Frozen fixed points are stored in the same way as formulas in Gamma
   (i.e., with a key-value pair of index and formula), causing minor
   changes in init*Expert.

 - decideLClerk' takes into account the new indexing structure used for
   lemmas.

As indicated, changes are local and may be incomplete and/or brittle,
and meticulous cleanup will be necessary once the design has been
fortified.