GHR/kernel: init*Expert restricts by index
The gap that was introduced during the initial design stage, by which
the index to be read in the init*Expert rules was not fully defined, is
now resolved.

Both experts are extended to return the missing index, which is
compounded to make a full read from the (now indexed) frozen zone.
If this functionality is lacking, naming will not solve the ambiguity
problem.

Adjustments to the FPC must follow.