GHR: init*Expert extended in FPC
An index is generated from the naming structure associated to the
fixed point under consideration.

Note that we are leaving special cases (grounded on a constant,
irrespective of further branches) untreated for now. Also, given that
full indexes include a numeric (internal) and a symbolic (external)
component, a don't care numeric index is being added by convention,
though this should be specified, and possible unified in automation,
somewhere else.

Documentation should follow if needed.