Refactoring of naming structure manipulators
A probably not optimal, but certainly more modular structure is
implemented in clerks and experts, thus splitting treatment by
certificate types and treatment by naming structure manipulations.
Readability and intent are improved and repetition minimized rather
reasonably.

In addition, helper predicates for the naming structures are
generalized and renamed for greater clarity.

Missing cases are added for all relevant clerks and experts regarding
the fallback case of propagation of "leaflike" naming structures.

The one significant abscence is an integral treatment for induction,
which does prevent application to most proofs we have registered to
date, given that they usually rely on initial application of an
inductive rule.