GHR: immediate induction and related corrections
In order to offer proper support for induction, some significant tweaks
are introduced. First and chief in importance, the naming structure
manipulator associated to indClerk' must construct an appropriate
naming structure for the immediate invariant, based on the information
(hypothesis names, etc.) given from the top and already in our power.
This requires the tricky strategy of knowing the shape of these
invariants and replicating them... even though they are the kernel's
business! Thus, beware of changes in either, especially in the kernel,
which must be manually propagated for now. The better strategy seems to
be peeping or leaking of read-only information from the kernel.

In addition to this, immediate induction certificates must facilitate
additional information. Namely, because the fixed point that will be
developed may contain other fixed points, and not only the recursive
placeholder, a matching naming structure for the definition of the fixed
point must be given with names to those fixed points that will recur
later on. The naming structure for the immediate invariant will be
grafted to a placeholder for the recursive fixed point, which is given
by "(name "S")" as our current external interface (this is hardcoded in
the helper predicate, consider generalizing; also note the negative goal
and ensure correctness!!). The naming structure is added as a parameter
to the immediate induction certificate.

Minor extensions are made to the following naming manipulators:

 - andClerk

 - storeLClerk

 - decideLClerk

The debugging predicates are extended.

Treatment of induction with an explicit invariant must follow.