GHR: Most examples without explicit invariants
Almost all examples not involving explicit supply of an inductive
invariant are revised to work with the new certificates in "backwards
compatibility mode" (that is to say, no additional constraints are
introduced, and all frozen fixed points are directed to a single,
trivial pool, as was the case before implementing GHR). The
certificates are more verbose, but no changes should be appreciable.

The one exception is the proof of commutativity of multiplication,
which (if I recall correctly) did finish, but only after much
backtracking and failed attempts.