Lemma guessing: Finishing some simple proofs
If every hypothesis and conclusion belongs in the same bucket (i.e. a
common name category) or structures are clear and logic variables given,
it is easy to leave holes at the end of proofs and search for solutions.
Leaving holes in the middle should be possible only if we use big
buckets, because otherwise we won't be able to tell where to pick our
hipotheses and goals from!

The next stage is to derive these naming structures without explicit
user interaction, and hopefully output them for information as well.