Lemma guessing: constructive naming structures
The FPC is extended to allow guessing of naming structures. This is
necessary whenever we want to mix trivial lemma guessing ("which lemma
am I using?") and structure guessing ("what are my hypotheses and
conclusions?"). As mentioned in an earlier commit, lemma guessing with
the standing naming structures could only work if a fallback case for
the "bucket" where all of the hypotheses and conclusions that were
needed were included. In practice, this meant going for an all-or-
nothing approach.

Because we cannot, in general, ensure the names of the hypotheses for
the lemmas we are guessing, or even the lemmas themselves (how many
hypotheses?, how many conclusions?, and so on), we need to do something
different. A number of technical solutions are available that avoid
changes to the kernel. The two main ones that were considered were:

 - (As done here,) generate the naming structure dynamically according
   to the lemma, and do so without changing the way lemmas are stored.
   That is to say, do not request any further information to the user.

  - Conversely, let the user supply a naming structure template for
    each lemma, and generate for each application a new instance with
    logic variables to be filled out during the process.

The only way to proceed with the latter option without touching the
kernel is to complicate the indexes, which are the single point of
interaction (message passing) between the kernel and the FPC, together
with term instantiation. This seems rather indecorous, though it could
be implemented with minimal changes to the certificates to be checked
(of course, the lemma libraries would absorb the impact of this
modification). Though not without interest, refining and outputting
certificates will be of interest, and not doing so will prevent any
interactive use of the certificate: in fact, those instantiated
hypotheses, the new conclusions... will have private names, that the
user will not be able to query, or use in subsequent derivations.

Here we extend the FPC with a naming structure constructor that
indicates that, from the current point onwards, we shall need to make
guesses. Originally it was thought to be used exclusively at the top
level of a lemma application to signify that hypotheses will be left
to the checker to find. However, more flexible uses (e.g. guess from
a certain point onwards) may be possible, though have not been tested.

The basic idea of the implementation is as follows:

 - Name manipulator predicates must extend their case analysis to cover
   guessing. They do so by replicating the structure that would be found
   if the structure were defined (essentially, splits), introducing at
   each point new guessing placeholders for further refinement.

 - freezeLClerk and initRExpert are extended to accommodate the new
   discipline, i.e. by realizing whether a concrete 'name' is given (as
   was always the case before, though this was not made explicit) or
   a guess must be made in its stead, in which case the guessing branch
   comes to an end and the choice is passed, unconstrained, to the
   kernel.

The resulting concrete (guessed) naming structure is returned in a
verbose but compactable form. The implementation plays with logic variables
in a rather smart way, refining the guesses at each step a little more and
passing the new (open, logic variable) refinements as new guesses. The
organization of the code means that there are never any open ends, and
everything is instantiated in order. (Again, this is a high-risk area,
and attention should be paid to its correctness and clarity.)

Further notes:

 - The new case analysis, both in manipulators and init/freeze, should
   be reviewed for compactness and style.

 - Most examples have been re-checked, but the combination of single
   freeze indexes and open lemmas leaves room for significant choice,
   and though I am inclined to think that this always works, it should be
   verified.