GHR: Support for explicit invariants
The administrative FPC is extended to once again accept explicit
induction certificates. In order to do this, the corresponding
constructor is given two naming structures as parameters: one detailing
the naming conventions for the explicitly supplied invariant; the other,
as in the case of immediate induction, providing names and the location
of the invariant in the fixed point that will be induced upon.

The treatment essentially mimics that implemented for immediate
induction, where some of the naming structures and parameters are given
explicitly, and both branches of the induction rule must be elaborated.
The code reflects this relation and could possibly be compacted further
(note also that the same warnings and caveats that are valid in
immediate induction remain current here).

Debugging functions and existing examples relying on explicit induction
are likewise updated.