Coinduction: first examples
Battery based on finite automata (cf. Abella tutorial). The examples
anticipate the at least temporary re-use of induction certificates for
quick prototying of coinduction.

An interesting point will be the failure to retrieve the name of the
least fixed point, sim, based on the fixed point formula itself, due to
a curious unification error that seems to depend on the nesting of
least fixed points in its structure. As of yet, no specific cause has
been located.