Coinduction: first changes to FPC/kernel
Incomplete support for coinduction is introduced through the following:

 - Definition of coindClerk and helper predicate for naming structures.

 - Definition of the freezeLClerk/initRExpert pair.

In the latter, we note that given the single frozen spot on the right,
no indexes are necessary, and this change is temptatively introduced in
the kernel, to be subsequently confirmed and all interfaces cleaned up.

Currently, it must be noted that greatest fixed points present
difficulties to be tackled in subsequent commits. Namely, the argument
vector representation of generic fixed points generates a string of
existentials during the parameter parsing stage. When manipulating
greatest fixed points on the right, this will result in an
administrative mini-bipole that we are currently not prepared to handle,
given the tenet that, once focused on the right, the focus shall never
be released. We do need, however, to do exactly that, if our treatment
of fixed points under this convenient encoding (and, possibly, any
other compact treatment) is to be preserved.