[Bug] Coinduction: pretty printing with example
Greatest fixed point lookup by name is adjusted in the logic debugging
functions to match that correctly used by least fixed points (I could
not quite recall the comparison had to be done a bit smartly, and the
mirror code for the greatest case was not up to date). Behold the danger
of ignoring DRY.

A trivial example of P -> P on greatest fixed points is added and checks
correctly with and without pretty printing under the current framework.