Coinduction: FPC extensions and first theorems
The administrative FPC is extended with the following clerks and
experts, which are needed for the new kinds of invariants and proof
procedures:

 - ttClerk

 - ttExpert

 - releaseRExpert

The main novelty is the (potential) need, in some instances, for a
release on a focused right sequent. I am not so sure of this anymore,
but for now a control parameter has been added to the limits exactly
for this purpose. So there is a real possibility that this bit will go
away.

Some examples based around automata are added, and here we can see some
potential limitations of the approach. First, reflexivity of sim works
out of the box (with the invariants supplied in Tac), but transitivity
requires that the "trivial branch" certificate be specified more fully.
(This we will do next.)

More troublesome is the failure to verify computational properties of
the greatest fixed point. The encoding has been adjusted to be the dual
of the least fixed point, but even so we run into problems that, in all
likelihood, arise from my compact multi-argument encoding. A fix would
be the best option, as I am reluctant to unfold them if can be at all
avoided.