Coinduction: Limitations of Gamma indexing
Adjusting the certificate to proceed with the proof of transitivity of
simulation reveals the next, and rather significant, problem.

By virtue of duality and the structure of the proof system, greatest
fixed points do not need a separate frozen bin, or even shared space in
the existing one: only one greatest fixed point can be frozen, and it
can never leave the right side of the sequent. So what, now, can we
match it with? The candidate fixed points are stored, indeed, in the
Gamma bin, and here is where the problem comes.

Consider the way we have indexed the frozen area, where indexes are of
no import except to restore naming structures: currently, all frozen
atoms will be given an opportunity to match the corresponding least
fixed point on the right.

However, with Gamma we wanted to be smarter and implemented a somewhat
limited, but so far effective, heuristic: namely, we numbered each
formula that went into the bin and recalled it exactly once and in the
same order. This helped with efficiency very significantly in the early
stages (how much of an advantage it is with current "lean" certificates
remains to be seen). Now with greatest fixed points, we may have a
number of hypotheses stored in Gamma and only one of them, not
necessarily the first, may match the point on the right.

A tentative proposal is to deprecate the indexing discipline and see
what happens: the present hurdle should be overcome, but what becomes of
the examples that were picking things efficiently up to now? Will there
be anything else that we can do?