Coinduction: Abstract properties and FPC support
To tackle some of the questions raised by the previous commit, the FPC
is extended as follows:

 - unfoldLExpert is defined by analogy with unfoldLClerk (sharing, at
   least for now, its limit counter). This enables treatment of
   greatest fixed points on the left in a more complete manner than was
   supported until now.

 - decideLClerk is generalized to select at any given decision point
   any one of the formulas stored in Gamma, as was done for lemmas
   before names were introduced as indexing and selection mechanisms in
   certificates. This means that more decisions are possible at each
   such point, but greater generality is obtained as a result. Whether
   this is strictly necessary at this point or not remains to be
   verified. (The code here can be simplified, as well.)

Simulation on automatas is furthermore generalized to prove reflexivity
and transitivity without resort to any one particular definition of the
step predicate. For now, we do this simply by (re-)defining the
predicate to falsehood, which guarantees immediate failure whenever its
structure is required, e.g. by unfolding, so that freezing becomes the
only option that can possibly yield results.

The question remains of whether the current model easily allows
computations on greatest fixed points to succeed, e.g. those based on
the simulation properties of specific transition systems (as defined by
the nefarious step predicate). Otherwise, results are promising.